Refactor server configuration logic into a separate module and update related tests
