+++ __tests__/app.test.js
@@ -1,30 +1,19 @@
 import request from 'supertest';
-import express from 'express';
-import { fetchAndParse } from '../fetchAndParse';
+import { configureServer } from '../configureServer'
 
-jest.mock('../fetchAndParse'); // Mock the fetchAndParse module
+// Automatically mock the fetchAndParse module
+jest.mock('../fetchAndParse', () => ({
+  fetchAndParse: jest.fn()
+}));
+const { fetchAndParse } = require('../fetchAndParse');
 
-const app = express();
-app.use(express.json());
+// Create the app using the configureServer function
+const app = configureServer(fetchAndParse);
 
-app.post('/parse', async (req, res) => {
-  try {
-    const { url, options } = req.body;
-    if (!url) {
-      return res.status(400).send('URL is required');
-    }
-
-    const parsedData = await fetchAndParse(url, options);
-    res.status(200).json(parsedData);
-  } catch (error) {
-    res.status(500).send('Error processing your request');
-  }
-});
-
-describe('POST /parse', () => {
+describe('POST /extracts', () => {
   it('should require a URL', async () => {
     const response = await request(app)
-      .post('/parse')
+      .post('/extracts')
       .send({});
 
     expect(response.statusCode).toBe(400);
@@ -35,7 +24,7 @@ describe('POST /parse', () => {
     fetchAndParse.mockResolvedValue({ data: 'mocked data' });
 
     const response = await request(app)
-      .post('/parse')
+      .post('/extracts')
       .send({ url: 'http://example.com', options: {} });
 
     expect(response.statusCode).toBe(200);
@@ -46,7 +35,7 @@ describe('POST /parse', () => {
     fetchAndParse.mockRejectedValue(new Error('mock error'));
 
     const response = await request(app)
-      .post('/parse')
+      .post('/extracts')
       .send({ url: 'http://example.com', options: {} });
 
     expect(response.statusCode).toBe(500);
