+++ app.ts
@@ -1,34 +1,8 @@
-import express from 'express';
-import { fetchAndParse } from './fetchAndParse'; // Import the fetch and parse module
+import { fetchAndParse } from './fetchAndParse';
+import { configureServer } from './configureServer'
 
-const app = express();
+const app = configureServer(fetchAndParse);
 const port = 3000;
-
-// Middlewares
-app.use(express.json());
-
-// Health Check Endpoint
-app.get('/', (req, res) => {
-  res.status(200).send('OK');
-});
-
-// Define the route for fetching and parsing a webpage
-app.post('/extracts', async (req, res) => {
-  try {
-    const { url, options } = req.body;
-    if (!url) {
-      return res.status(400).send('URL is required');
-    }
-
-    console.log(`Fetching and parsing ${url}`);
-
-    const parsedData = await fetchAndParse(url, options);
-    res.status(200).json(parsedData);
-  } catch (error) {
-    console.log(`error: ${error}`);
-    res.status(500).send('Error processing your request');
-  }
-});
 
 // Start the server
 const server = app.listen(port, () => {
